TT-1613: Switch hub-policy to fetch valid IDPs using LoA in addition to Transaction
Add transitional versions of state classes to be replaced to enable a zero-downtime deployment. It wil be possible for the app to consume old versions of the states present in the cache, but it will not produce any new instances of the old versions and wil use the new definitions instead. If an old version is retrieved from the cache, its instance is converted into the new definition for internal use.

Authors: @CharlesIC