TT-1613: Switch hub-policy to fetch valid IDPs using LoA in addition to Transaction
Add transitional versions of state classes to be replaced to enable a zero-downtime deployment. The app will be able to consume these new versions but will not create new instances of them just yet to avoid breaking the caching and to enable ZDT deployment. If an new version is retrieved from the cache, its instance is converted into the old definition for internal use.

Authors: @CharlesIC