TT-1613: Only store transitional versions of old state classes to enable a zero-downtime deployment.
In this stage, the app will only store instances of the new transitional state classes in the cache, so that the old instances won't be stored anymore, but will still be used internally. This will allow the old definitions to be altered, and then we will transition to using them again from the temporary transitional classes.

Authors: @CharlesIC