BAU: Port class from security-utils to support x509 certs in fed config
Due to changes to the way DeserializablePublicKeyConfiguration works,
fed-config DTOs can no longer use that class to get certificates.
Due to an outstanding bug, we cannot use the specific subtype so
this commit ports the relevant class (and merges the abstract super)
to perform the same behaviour locally in hub.

There is a plan to eventually migrate these DTOs to a database rather
than file backed system in which case this issue will be redundant

Authors: @andy-paine