BAU: Upgrade security utils
Due to changes to the way DeserializablePublicKeyConfiguration works,
fed-config DTOs can no longer use that class to get certificates.
Due to an outstanding bug, we cannot use the specific subtype so
this commit ports the relevant class (and merges the abstract super)
to perform the same behaviour locally in hub.
To allow the DTO getters to remain the same, certificates are new-ed
inside the getters. This requires the base Certificate class to be
comparable to prevent referencing issues so an equals and hashCode
method were added

There is a plan to eventually migrate these DTOs to a database rather
than file backed system in which case this issue will be redundant