ABD-35: Add a feature toggle to send events to a recording system
Add a configurable toggle to policy, saml soap proxy and saml proxy
to send or not to send events to a new recording system. The default is
not to send events to the recording system since the recording system is
not set up yet. Please note that this change does not affect Hub sending
events to Event Sink (i.e. Hub will continue to send events to Event
Sink even when the toggle is set to not to send events to the new
recording system).
Clean up code (e.g. break a very long line into multiple lines to make
it easier to read).

Authors: @adityapahuja