ABD-35: Upgrade to use latest version of rest-utils
to support the ability to send encrypted events to either a source queue
or a stub queue. If the source queue name is not present, encrypted
event messages will be sent to the stub queue. If the source queue is
present, encrypted event messages will be sent to the source queue. The
stub queue will just write the encrypted event message to the standard
output.

Authors: @adityapahuja