EID-430: Added validation for AuthnInstant in AuthnStatements
- UK Proxy Node assumes IDP Responses contain AuthnStatements with
  AuthnInstants set. We added validations for this
- Refactored DuplicateAssertionValidator to throw on validation failures
  rather than return a boolean. This will help us in testing since we
won't have to always explicitly mock it to return a value. The reason
why it was returning a boolean before was so consumer classes could
throw a different error based on validation result. The validator class
itself should just know which error to throw based on which set of
assertions it is validating

@minhngocd