refactor: rewrite Response generator tests
These tests were broken and were not testing the right things. Fixes
include:
- Testing that we are actually encryption assertions (and including
them)
- Testing that we are skipping signature when we don't want them (we
were actually signing before and then comparing with a message that had
a signature (not good!))
- Test that we are using the right features for legacy and profile
support (this wasn't being tested properly and the fact that we have
caching within saml-engine wasn't being caught)

`ConfigStubRule` was also behaving awkwardly so I've rewritten it using
wiremock, which should help us to migrate the rest of hub later.