HUB-103: Update to use latest version of utils-libs
Update to use latest version of verify-utils-libs which removes the
getClientTrustStoreConfiguration method from the TrustStoreConfiguration
interface. Client truststore is not used anymore for Certificate Chain
Validation for Hub or IDP trust stores. This is now all done by a filter in the
Metadata Resolver. Certificate Chain Validation is still done for the RP
truststore so the overall function needs to remain.

Co-authored-by: Rachel Smith <rachel.smith@digital.cabinet-office.gov.uk>