Log useful request / response attrs as JSON
Currently the ProtectiveMonitoringLogFormatter logs something which at
first glance may appear to be JSON. It has {} and keys/values separated
by : however don't be fooled. This is not JSON, there are no quotes "!.

This change places attributes which were part of the LogFormatter into
the MDC which is included in the Logstash output as JSON. Having
structured logs makes it easier to query them through ELK.

As this change doesn't alter the message logged it shouldn't impact any
current users of the logs (e.g. SecOps).