Remove boxed Boolean type from ProtMonLogging
Previously we were using a Boolean type to represent the following
 - null: no signature on the request/response
 - false: signature present but not valid
 - true: signature present and valid

This can be problematic as null checks need to be carefully performed.
These checks were missing in a035bb51a2f469022dbb25bd163e4d9a49d3ce8b

This commit replaces the usage of `null` with a separate isSigned
boolean. The change is only necessary for when processing Response
objects as they have this option of being unsigned (simple profile).