Introduce SignatureStatus enum
Thanks to @richardTowers for highlighting that using two booleans ended
up with a state which didn't make sense:

    isSigned && signatureIsValid
    isSigned && !signatureIsValid
    !isSigned && !signatureIsValid
    !isSigned && signatureIsValid // Doesn't make sense

This change introduces an enum for the top 3 states, VALID_SIGNATURE,
INVALID_SIGNATURE and NO_SIGNATURE respectively.

Now the "not signed, with a valid signature" state is no longer
possible.

Thinking about how the MDC is used in our logging platform this will
also allow for writing more efficient queries against a single field in
ElasticSearch.