bau: Fix failing tests when log level set to WARN
The tests were relying on a side effect of the console appender - it
calls getMDCPropertyMap() which populates the map lazily. If the log
level was set so the console appender did not fire, this method was
never called and the mock appender ended up with the wrong map.

To work around this we're reverting back to using a custom class instead
of a mock and explicitly adding the side effect of calling
getMDCPropertyMap so that the mock behaves more like a real appender and
the value we assert on at the end of the test is correct.

Co-authored-by: David King <david.king@digital.cabinet-office.gov.uk>