bau: Publish hub-saml to bintray
I've had to unroll publish.gradle into hub-saml and hub-saml-test
because I couldn't work out how to get the bintray plugin to apply
properly inside the configuration block. This way should be more
straightforward to understand, if at the cost of a bit of duplication.

This will allow people to build the things that depend on hub-saml (MSA
/ compliance-tool [if we ever open source it]) without access to
artifactory.