Revert "BAU: Remove hub SAML signing certificate from SAML Engine"
This reverts commit 736c581173a65e0a3fa244733c69699613f3ffd9.

Whilst we do not need the signing certificate for the regular Verify
federation, we do need it for communicating with EU countries via eIDAS.

The KeyInfo part of the SAML message needs the full X.509 certificate, not just
the public key.

Reintroduce the SAML signing certificate for now - we may be able to retrieve
it from SAML metadata in future.