Add prometheus support to Policy behind a feature flag
This adds PrometheusBundle from alphagov/gds_metrics_dropwizard.  This
adds configurable support for prometheus metrics to be exposed at
/prometheus/metrics on the admin port.  By default, this endpoint is
disabled; it is activated by adding `prometheusEnabled: true` to
policy.yml.

Policy was chosen arbitrarily as a proof of concept; there's no reason
we couldn't do the same for all apps.

In order for this to work, the gds-metrics-dropwizard dependency will
need to be added to artifactory.