BAU: Update to use ObjectMapper provided by DropWizard
Modules classes create a new instance of ObjectMapper that is not same
as the instance of ObjectMapper provided by DropWizard. This results in
logging timestamp not following the ISO8601DateFormat. The fix is to get
Modules classes to use the existing ObjectMapper provided by the
DropWizard instead of a new instance of ObjectMapper. This will allow
applications to log timestamp in the ISO8601DateFormat.

Authors: @adityapahuja