EID-951: Make ResponseFromCountryValidator allow non-successful auth responses from Countries to contain unencrypted assertions
In the eIDAS world, authn responses are allowed to have unencrypted assertions if they don't have the "Success" status. Some countries include such assertions in their Cancel responses but currently ResponseFromCountryValidator throws and exception and we can't handle this type of failure in a graceful manner, so the users see the "Service unavailable" error page. If we allow these to pass, then the users will see the new error page with a better message and a link to retry the journey.

Some of the countries that send such responses are Germany and any country that uses the CEF Reference Node (e.g. Belgium and Estonia).

Author: @CharlesIC