EID-951: Add the new CountryAuthenticationStatus enum so we can handle authn responses specific to eIDAS and Countries.
Currently, we're re-using IdpIdaStatus enum to handle authn responses from Countries. This enum is not well-fitted for eIDAS, though, because the SAML responses from IDPs and Countries and their properties vary significantly for any journey that's not a success. The Country status definitions will allow us to handle eIDAS-specific responses better and customise them without affecting the Verify flow.

Author: @CharlesIC