HUB-307: Add tests for when there are missing certs & multiple certs
Added tests to check that we have correct behaviour when there are no certs
and when the metadata contains multiple certs. (e.g. during a rotation)

Fixed extractor to account for new tests.

Added runtime check in extractor to verify that the returned cert's public
key is the same as the one we passed as a parameter.