EID-1004: Fix transition between Verify and eIDAS journeys (part 2/2)
Currently, if a user selects an IDP without completing verification and clicks the back button to the journey picker page and then attempts an eIDAS journey, they receive a "service unavailable" error page as Policy doesn't allow transitions from IdpSelected to EidasCountrySelected states. The solution to this is to allow the frontend to ask Policy to restart the journey back to SessionStartedState when a user goes back to the journey picker page (/prove-identity) and selects a an eIDAS journey. We already allow this for eIDAS journeys when a user goes back and selects a Verify journey.

This piece of work enables transitions from a Verify journey to an eIDAS journey. Transitions on the opposite direction were enabled in part 1.

Summary of changes:
  - Remove the no longer used restart-eidas-journey references as they have been generalised to restart-journey
  - Make the IdpSelected state and controller implement the RestartJourney interface to give the IdpSelectedState the ability to be reset to SessionStartedState
  - Add tests to reflect the above changes

Author: @CharlesIC