EID-1004: Fix transition between Verify and eIDAS journeys
Currently, if a user selects a country or an IDP without completing verification and clicks the back button to the journey picker page and then attempts a different type of journey, they receive a "service unavailable" error page as Policy doesn't allow a transition between IdpSelected and EidasCountrySelected states. The solution to this is to allow the frontend to ask Policy to restart the journey back to SessionStartedState when a user goes back to the journey picker page (/prove-identity) and selects a different type of journey. We already allow this for some eIDAS states. This piece of work makes the interfaces more general so that we can apply the same to the IdpSelected state later. This commit only affects eIDAS -> Verify transitions. Transitions in the opposite direction will be enabled in a later release.

Summary of changes:
  - Rename EidasUnsuccessfulJourney state and controller interfaces to RestartJourneyState to make them more general
  - Make the EidasCountrySelected state and controller implement the above interface to give the EidasCountrySelectedState to be reset to SessionStartedState
  - Add new more general URL to Policy that frontend can call to restart a journey. The old URL will be removed in a later release after a frontend release
  - Add tests to reflect the above changes

Author: @CharlesIC