EID-1015: Unhandled exception when user fails MW auth and goes back and re-submits form
Currently, if a user fails authentication with the Middleware and and goes back and retries, Policy will throw an INVALID_STATE exception. SAML Proxy will pick that up and in turn return an HTTP response. Frontend expects a JSON body, however, and cannot parse the response. The JSON library code throws an exception that may be handled at the top level. In prod users get correctly redirected to the "Something went wrong page" but we see an entry in Sentry every time with the message "Couldn't parse JSON" instead of the actual exception message. In non-prod environments, devs see the Ruby error page.

Summary of changes:
  - Add a line to AbstractContextExceptionMapper in SAML Proxy to make sure that it returns a JSON body that Frontend can handle.
  - Add /EidasResponse/POST path to the list of paths which we don't expect to have context. The regular IDP response path was in that list but the eIDAS path wasn't and that caused INVALID_STATE exceptions for eIDAS journeys to not be handled properly.
  - Add exception message to the JSON body when handling SamlProxyApplicationException

If Policy throws INVALID_STATE exception from now on, devs will see the generic Verify "Something went wrong" page and the message in Sentry will reflect the actual error.

Author: @CharlesIC