EID-1015: Unhandled exception when user fails MW auth and goes back and re-submits form
The Middleware design permits the unusual behaviour that allows an authn response from the MW to be POSTed multiple times to the Hub for one authn request sent to it. This happens when a user fails authentication, presses the back button and refreshes the page which causes the browser to re-submit the form with the authn response. Occasionally the user would go back to the MW app itself and submit another response for the same request. At the moment, when this happens the user would see the "something went wrong" page and we'd log an exception. A better way to handle this, as this is a known issue, is to send the user back to the page they came from, i.e. the eIDAS error page. This will also ensure we only process one response for each request.

Summary of changes:
  - When handling a response from a country, check if the session is in EidasAuthnFailedErrorState. If it is, return a FAILED response which will allow frontend to redirect the user to the eIDAS error page
  - Add unit and integration tests for the above

Author: @CharlesIC