EID-1072 Bump saml-libs to version 169
It pulls in a few new things.

An update to EidasMetadataResolverRepository to throw an exception
instead of an error if it encounters certificate that's expired. This is
useful in saml-engine and saml-proxy.

MetadataResolverBundle has been updated to generate a dropwizard
healthcheck for the metadata, rather than the app itself having to do
it. This meant the healtcheck being created in saml-engine could be
removed.

It allows optional metadata bundling, which required an update to the
config class for saml-engine as well as saml-soap-proxy to support
Optionals being returned by `getMetadataConfiguration`. The metadata is
always not null in these cases to impact is minimal.