AUTO-146: Decouple PrometheusClientService from ConfigDataBootstrap
Creates CertificateExpiryDateCheckService class which is responsible for updating all certificates' expiry date.
Creates PrometheusClientServiceConfiguration which allows us to configure the following two configuration items for a service.

1. initialDelay - the time to delay first execution. It is set to 10 second by default.
2. delay - the delay between the termination of one execution and the commencement of the next execution. It is set to 15 minutes by default.

Updates PrometheusClientService to let DropWizard to manage a CertificateExpiryDateCheckService in its ScheduledExecutorService using configurable initialDelay and delay.
Removes a dependency between PrometheusClientService and ConfigDataBootstrap. There is no need for PrometheusClientService to wait for ConfigDataBootstrap since CertificateExpiryDateCheckService is part of ScheduledExecutorService and it can wait for ConfigDataBootstrap to complete its task using the initialDelay.
Updates Certificate Expiry Date Check metrics to include a timestamp field required for show that the metrics are not stale.
Updates Prometheus Metrics Integration Test to test that it can fetch regularly updated certificates metrics.

Author: @adityapahuja