BAU Log IDP response action 'OTHER' created
When we receive an authn response from an IDP we generate a response
action which is returned to the frontend.

We've had a few users and on the 'failed' page on the frontend. This
will happen if we return a response action of 'OTHER'.

Currently we have no insight into when this happens. This will help.