AUTO-146: Move last updated from labels to a separate metric
Every unique combination of key-value label pairs represents a new time series which can dramatically increase the amount of data stored. We should not use labels to store dimensions with high cardinality (many different label values) such as timestamp or other unbounded sets of values. This commit updates to have pair of metrics (one for expiry date or OCSP revocation status and another one for last updated (timestamp)). The last updated metric is used to indicate whether expiry date metric or OCSP revocation status metric is stale or not. It will be used to alert support team that the metrics are out of date and the team will fix the issue. This commit also adds a serial number to certificate metrics since metadata certificate metrics already have serial numbers.

Author: @adityapahuja