AUTO-146: Update OCSP last success timestamp
This commit changes from updating OCSP revocation status last update to
OCSP last success timestamp. In this way, the alert will be "have we
had a successful OCSP check for each certificate in the last N hours?"
rather than "was the most recent OCSP check successful?". This change is
needed because support team do not want to get alerted when OCSP server
is unavailable.

Author: @adityapahuja