AUTO-152: custom collector for info metric
This introduces a custom collector [1] for the _info metric.  The idea
is that we only want to store one metric label set per matching
service.  Previously we achived this with an info_last_updated metric
but this led to complex promql expressions to select only the most
recent info metric

This commit introduces MatchingServiceInfoMetric, a custom collector
which stores the latest information for each matching service, stored
in a ConcurrentMap by URL of the matching service.  (It needs to be a
ConcurrentMap because we might get scraped at the same time as we're
performing a health check.)

[1]: https://github.com/prometheus/client_java#custom-collectors

Co-authored-by: adityapahuja <aditya.pahuja@digital.cabinet-office.gov.uk>