BAU: Fix Cycle3DataInputCancelledState state serialization
The field type for relayState is Optional but it was being serialized to `null` when absent. This meant that when we tried to deserialize it, the Optional was set to null (since it is a reference type). This commit fixes it so that the constructor takes just a String and wraps it in Optional if it is null.

Author: @adityapahuja