expose metrics for trust store certificates
This exposes metrics for certificates in truststores, so that we can
see what's actually configured and deployed in each environment.
Metrics are added to each of saml-proxy, saml-soap-proxy and
saml-engine.

This uses the TrustStoreMetrics class introduced in
alphagov/verify-utils-libs#29.

SamlProxyApplicationTest and SamlSoapProxyApplicationTest failed with
a NullPointerException but that's because they call
Saml[Soap]ProxyApplication.run() with incomplete configuration.  In
normal operation, we validate that Saml[Soap]ProxyConfiguration is
@NotNull so we don't need to add a null check.

Furthermore, I don't think these are useful tests - they're just
checking that if we don't set
Saml[Soap]ProxyConfiguration.eventEmitterConfiguration, then it will be
null.  So I deleted them.