BAU Upgrade dropwizard to 1.3.9
1.3.5 was susceptible to a range of CVE's which are fixed by [1.3.9](https://github.com/dropwizard/dropwizard/blob/master/docs/source/about/release-notes.rst#v139-feb-24-2019).

One of the dependencies that was bumped by this patch was Apache
HttpClient to v4.5.7. 4.5.7 introduced an issue where double forward
slashes in URIs would be normalised to just one.

This caused a regression in SAML engine as we use entityIds, which are a
URL, as part of the URL when fetching their encryption keys from config.

This issue is discussed in detail [here](https://issues.apache.org/jira/browse/HTTPCLIENT-1968)
and should be fixed in future (potentially late March or early April).