BAU: Increase MSA healthcheck timeout
MSA timeout is 50 seconds whereas MSA healthcheck timeout is 2 seconds.
There are 281 timeouts in MSA healthcheck in production over the last 7
days via Kibana. This commit updates the MSA healthcheck timeout to 50
seconds which is same as MSA timeout. This change will reduce the number
of timeouts occurred in MSA healthcheck.

Author: @adityapahuja