EID-1362: Use hashed PID for eIDAS logging
The hash of user attributes logged by saml engine gets compared with a
hash of user attributes in the proxy node. The PID used to produce this
hash in the proxy node has already itself been hashed by the VSP so we
need to replicate this behaviour in the EidasAttributesLogger.