BAU Create snyk scripts
These are almost directly copied from the scripts used in the
proxy-node.

The main difference is the names of the configurations to test. Also
changed is the gradle sub project being used.

The scripts now also move the Gemfile and Gemfile.lock in the root of
the project to a different namespace while the tests run. Without this
snyk will just test the dependencies in there rather than the gradle
dependencies.