BAU Bump Guice version in common config
This fixes CVE-2018-10237, a medium severity vulnerability.

Guice 4.0 was pulling in version 16.0 of Guava which was vulnerable.

Unfortunately from Guice 4.2.1, Guice pulls in Guava 25.1-android which
is incompatible with our stuff. Guava release two variants `*-android`
and `*-jre`. The recommendation is to explicitly declare a dependency on
the `-jre` variant of equal or greater version number. See (here for a
pretty good explanation.)[https://github.com/google/guice/issues/1207]
This has been done.