BAU Update travis.yml to run snyk scripts
This adds the SNYK_TOKEN env variable as an encrypted variable. The
token used is my token.

It installs the CLI before the projects dependencies are installed in
the "before_install" block.

It runs test.sh after running the default check task, as before.

It runs monitor.sh if the "script" block was successful.