PwUqib5x: Increase MSA SOAP response log level
MSA errors should come back with stack traces. By logging them at an
`info` level, they are not being surfaced clearly in monitoring
tools. The errors that are thrown as a result don't have the MSA
stack trace so it is not always clear what the problem was. This
should help.

Co-authored-by: Brendan Butler <brendan.butler@digital.cabinet-office.gov.uk>