EID-1422 Use Verify's fork of Splunk logging library
We've forked Splunk's java logging library so that we can add the
ability to configure the proxy to use. The library uses an Apache http
client under the hood, and it doesn't use the java system properties by
default like the Jersey clients the apps use. The library offers no way
to configure the proxy; this is what was added by our fork. Hopefully we
can get the changes merged upstream.