l4zBlwLF: Add ConfigRepository layer
Add a layer between Transaction and Matching Service Resources and their
respective LocalConfigRepository. At the moment this is a simple
passthrough to the LocalConfigRepository, but this will be the place
where merging local and remote configurations will happen.

The existing MatchingServiceAdapterService was a more complex layer for
the MatchingService config. However the implementation did not do that
the names and tests suggested that it was trying to do. This has been
removed and replaced with the simpler ManagedEntityConfigRepository used
by both TransactionResource and MatchingServiceResource.