port AttributeQuery request Timer to prometheus
We have a Dropwizard Metrics Timer to record times taken for
AttributeQuery requests (ie SOAP requests to matching services).
However they are somewhat fiddly to deal with in prometheus.

This creates a prometheus Summary to capture this data.  A big
advantage is that the matching service URI will be available as a
label rather than a weirdly formatted part of the metric name, so we
can more easily do queries for things like "show me the slowest 3
matching service URIs".

I picked the quantiles to be a useful subset of the current quantiles
on the dropwizard Timer.