l4zBlwLF: Add Adobe S3Mock
This commit adds and and wires up Adobe's S3Mock for integration tests.
At this stage the `SelfServiceIntegrationTest` does not exercise code
that performs S3Integration, just wires up the mock AmazonS3 client
object so that it is availible. The next step is to add the code that
merges local and remote config and extend the integration test to
exercise that.