0BFh8hBN: Use enabled flag in TrustStore Config
We want to be able to turn off validating certificate chains with our
root trust store for RPs (and their MSAs). We have previously added an
`enabled` flag to the `ClientTrustStoreConfiguration` in
`verify-utils-libs`.

This commit:

Updates `TrustStoreForCertificateProvider` to return `Optional<KeyStore>`
from `getTrustStoreFor`.

Uses the new `enabled` flag in `ClientTrustStoreConfiguration` to
determine whether or not to return the configured trust store.

Updates `ConfigServiceKeyStore` to validate using the trust store only
if it is present. i.e. the Optional returned from `getTrustStoreFor` is
not empty.

(based closely on a recent saml-engine commit by @kerrr)