BAU: Kill the X509CertificateConfiguration class
We feel this class is doing more harm than good.  It only adds headers and strips whitespace (to create a so-called full certificate), and there is only one place where that format is needed, so the formatting may as well be performed there instead; this also means we no longer need code for undoing that formatting.  Making this change should help simplify the setup of some tests we want to add soon.

Co-authored-by: Kerr Rainey <kerr.rainey@digital.cabinet-office.gov.uk>