BAU: Update deserialization, merge Certificate
Remove X509CertificateConfiguration class. We feel this class is doing
more harm than good. It only adds headers and strips whitespace (to
create a so-called full certificate), and there is only one place where
that format is needed, so the formatting may as well be performed there
instead; this also means we no longer need code for undoing that
formatting. Making this change should help simplify the setup of some
tests we want to add soon.

The deserialization of `MatchingServiceConfig` and `TransactionConfig`
needs to be able to handle values for certificates being raw Strings and
being nested in hash with name "x509". We have enabled this by managing
the JsonNode directly.

The Certificate and CertificateDetail classes did not have clear roles
and responsibilities, blending into each other, CertificateDetail
composing certificate with some extra values. This has been flattened
with those values merged into Certificate

So we now have:

- Simple Strings for base64 encoded representations of certificates in config classes.
- The Certificate class that gives us a richer representation of a certificate with added metadata of issuer entity id, usage, etc. for general use within the Config Service.
- Java.security X509Certificate as the actual certificate.
- Review by commits. The first commit removes X509CertificateConfiguration, the second merges Certificate and CertificateDetail. The second commit is quite a big refactor, but with minimal changes that are not a result of the merge.

co-authored-by: Alan Carter <alan.carter@digital.cabinet-office.gov>