8HRTuUnq: Add CertificateOrigin
We need to track the origin of a Certificate, so we can determine what Validation it requires. We have added a new enumeration for CertificateOrigin and added it to classes that extend `CertificateConfigurable`.
The values set on these are in turn propagated onto the `Certificate` values returned from them. The critical place for setting this value is when overriding with remote certificates from Self Service.