8HRTuUnq: Update Certificate for Optional cert
The `Certificate` class still needs to be returned from `CertificateConfigurable` instances that do not have a certificate set on them. Therefore we need the underlying certificate wrapped by `Certificate` to be optional.
Updated CertificateValidityChecker to handle Optional x509 in `Certificate` and refactored for clarity.
Refactored CertificatesResource to handle Optional x509 in `Certificate`.
Fixed up tests.