EID-1608 Determine if assertions are unsigned (WIP)
We need to use saml-engine to determine whether the assertions are signed or not.

The original saml actually doesn't need passing back as it should be present in some form in handleEidasResponsePost

This commit is guilty of feature creep as it contains code that deals with decrypting the symmetric keys which is something for another story.

For this story, there just needs to be a property returned that says whether assertions were signed or not and then the original saml is sent instead of the re-signed message.