EID-1608 Pass original saml message if assertions are unsigned
In order to maintain trust when a country sends unsigned assertions, we need to provide the original saml message signed by the country so that the service can see that what they've received is what the country that sent it, sent.

To that end we need to determine whether message contains any unsigned assertions at the point it's translated.  As if it does contain unsinged assertions, we send back a areAssertionsUnsigned: true in the json and end point that received the inboundResponseFromCountry can decide what to do with the original message (of which it has a copy). This is just an added property in the response - nothing else changes in that respect.