Test JSON Serialization for Redis
Policy now uses Redis and serializes `State` data using JSON. We need to check that we don't make changes to the JSON serialization accidentally. Changes to the serialization that are unintended could cause ZDD problems during deserialization.

These tests check that the we are serializing to the current known format.