If assertions are unsigned, we need to pass the original saml and also, symmetric encryption keys encrypted for the RP (MSA/VSP).
The main change is to make use of the changes in verify-saml-libs from [EID-1609](https://github.com/alphagov/verify-saml-libs/pull/86).  CountryAuthnResponseTranslatorService#translate returns an InboundResponseFromCountry that contains the original SAML request (for simplicity), a boolean stating whether or not the assertions are unsigned and conditional on that, re-encrypted symetric keys so the RP/VSP/MSA can read the encrypted assertions in the original SAML message.