Remove more guava optionals
Fixed up serialization tests. We are not changing the json serialization
that we now use, so there shouldn't be any issues.

We could probably remove these tests as we don't use java serialization
anymore. Leaving them there for the time being so that we have to think
about serializing when making changes.