BAU Add logging to check status mapping logic
We have a theory that AuthenticationFailed statuses are not being mapped
correctly. If an IDP returns anything for status detail, the status
won't be matched. As this is the only status with a status of Responder
and a sub-status of Authn_failed, we shouldn't care too much about the
detail value.