BAU Don't check detail for AuthenticationFailed status
We've recently seen an increase in errors being thrown when an IDP
returns an AuthenticationFailed status. This is due to a recent change
in the logic for mapping the status from the SAML to a Java object.

This status has a top-level status of Responder and a sub-status of
Authn_failed. The recent changes meant that we failed to match this
status when mapping if the IDP also included a status detail value (we
_only_ matched if there was no detail).

We added some logging to check if IDP's were including detail values for
this status and it turns out they were.

As this status is the only status with this combination of top-level and
sub-status, we shouldn't care what the detail is and should match
regardless.