BAU: Update mockito version
Builds of `hub-saml` were failing in some environments when the build target was java 11. This was due to an old version of mockito that was not compatible with java 11 in some circumstances.

We are transitively dependant on Mockito via Junit 4, but we are already on the latest version of the 4 branch.

Forced upgraded to latest mockito 2.xx.xx version.