EID-1666 Update saml-engine to support unsigned assertions
This commit does a couple of things.

When a response from a country is POSTed to the hub for the first time
it's details are stored in a state object by policy. Policy calls SAML
engine with a saml response DTO, and saml engine parses it and returns
an `InboundResponseFromCountry` dto. This inbound resonse is used to
create the state object in policy.

This updates saml-engine to check if the assertions in the saml are
unsigned, and if not, include a `CountrySignedResponseContainer` in the
`InboundResponseFromCountry` dto. This includes the original saml
response from the country and the encrypted assertion keys required by
the service to decrypt the assertions. These details are added to the
state in policy in another commit.

Later in the journey, the frontend calls the hub again, to retrieve the
details of the response store in the session state in policy. Policy
then calls saml-engine to create a signed saml Response to return to the
services. This commit updates saml-engine to support the case where a
countries assertions are unsigned.

saml-engine will return a wrapper Response, with the original country
saml-response as an attribute value, and the encrypted assertion keys as
another attribute value, within the assertion. These attributes are
newly created extensions in saml-lib (see a separate PR for these). Once
the Response has been generated, it's signed as normal, but the
assertions are not encrypted. This is because the sensitive information
is _already_ encrypted for the service, and so can be left as data in
the assertions.

The Response is then returned to Policy, and back to the frontend, and
finally to the user via a form. This approach of saml-in-saml allows us
to not break the api of the VSP, which will be receiving these
Responses. It's worth noting that this new flow is only the case for
non-matching journeys. Matching journeys will handle the unsigned
assertions in a separate flow which includes sending messages to the
services MSA.